refactor utils.file
