+++ app/src/main/java/com/study/thesuperiorstanislav/edaapp/main/MainFragment.kt
@@ -15,9 +15,8 @@ import android.view.ViewGroup
 
 import com.study.thesuperiorstanislav.edaapp.R
 import com.study.thesuperiorstanislav.edaapp.UseCase
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.Circuit
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.Element
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.Net
+import com.study.thesuperiorstanislav.edaapp.utils.file.AllegroFile
+import com.study.thesuperiorstanislav.edaapp.utils.file.Calay90File
 import kotlinx.android.synthetic.main.fragment_main.*
 import java.io.BufferedReader
 import java.io.IOException
@@ -151,113 +150,11 @@ class MainFragment : Fragment(), MainContract.View {
         val reader = BufferedReader(InputStreamReader(inputStream))
         val line = reader.readLine()
         if (line == "\$PACKAGES") {
-            readAllegro(reader)
+            presenter?.cacheCircuit(AllegroFile.read(reader))
         } else {
-            readCalay90(line, reader)
+            presenter?.cacheCircuit(Calay90File.read(line, reader))
         }
     }
-
-    private fun readAllegro(reader: BufferedReader) {
-
-        val listElements = mutableListOf<Element>()
-        val listNets = mutableListOf<Net>()
-        val listPins = mutableListOf<String>()
-
-        var line = reader.readLine()
-        while (line != "\$NETS") {
-            line = reader.readLine()
-        }
-        line = reader.readLine()
-        var lastNet = Net("")
-        var isNetOver = true
-        while (line != "\$END") {
-            val splitLine = line
-                    .split(" ")
-                    .asSequence()
-                    .filter { it != "" }
-                    .toMutableList()
-
-            if (isNetOver) {
-                listNets.add(Net(splitLine[0]
-                        .replace(";", "")))
-                lastNet = listNets.last()
-                splitLine.remove(listNets.last().toString() + ";")
-            }
-
-            isNetOver = line.last() != ','
-
-            splitLine.forEach { str ->
-
-                val splitIt = str
-                        .replace(",", "")
-                        .split(".")
-
-                if (!listElements.contains(Element(splitIt.first()))) {
-                    listElements.add(Element(splitIt.first()))
-                }
-
-                listElements.find { it == Element(splitIt.first()) }
-                        ?.setPin(splitIt.last().toInt() - 1, str)
-                lastNet.addPin(str)
-                listPins.add(str)
-            }
-            line = reader.readLine()
-        }
-
-        presenter?.cacheCircuit(Circuit(listElements,listNets,listPins))
-
-    }
-
-    private fun readCalay90(firstLine: String?, reader: BufferedReader) {
-        val listElements = mutableListOf<Element>()
-        val listNets = mutableListOf<Net>()
-        val listPins = mutableListOf<String>()
-
-        var line = firstLine
-        var lastNet = Net("")
-        var isNetOver = true
-        while (line != null) {
-            val splitLine = line
-                    .split(" ")
-                    .asSequence()
-                    .filter { it != "" }
-                    .toMutableList()
-
-            if (isNetOver) {
-                listNets.add(Net(splitLine[0]))
-                lastNet = listNets.last()
-                splitLine.remove(listNets.last().toString())
-            }
-
-            isNetOver = line.last() == ';'
-
-            splitLine.forEach { str ->
-
-                val splitIt = str
-                        .replace(",", "")
-                        .replace(";", "")
-                        .replace("(", "")
-                        .replace(")", "")
-                        .split("'")
-
-                if (!listElements.contains(Element(splitIt.first()))) {
-                    listElements.add(Element(splitIt.first()))
-                }
-
-                val strPin = str.replace(";","")
-
-                listElements.find { it == Element(splitIt.first()) }
-                        ?.setPin(splitIt.last().toInt() - 1, strPin)
-                lastNet.addPin(strPin)
-                listPins.add(strPin)
-            }
-            line = reader.readLine()
-        }
-
-        presenter?.cacheCircuit(Circuit(listElements,listNets,listPins))
-    }
-
-
 
     private fun performFileSearch() {
         val intent = Intent(Intent.ACTION_OPEN_DOCUMENT)
